Final Test

+++ BoardProject/src/controller/action/ActionImgUpload.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.oreilly.servlet.MultipartRequest;
 import com.oreilly.servlet.multipart.DefaultFileRenamePolicy;
@@ -31,37 +32,44 @@ public class ActionImgUpload implements Action {
 		// 최대 16Mbyte 고정
 		int maxSize = 16 * 1024 * 1024; 
 		MultipartRequest multi;
-
-
+		
+		
+		
 		try {
 
 			multi = new MultipartRequest(request, filePath, maxSize, encoding,	// 파일 생성	
-					new DefaultFileRenamePolicy()); // 동일이름 자동 리네임		
-			userVO.setIconId((String)multi.getFilesystemName("filename")); // 생성된 파일의 이름 가져오기
-			userVO.setUserNum(Integer.parseInt(multi.getParameter("usernum")));
+					new DefaultFileRenamePolicy()); // 동일이름 자동 리네임	
+			int userNum=Integer.parseInt(multi.getParameter("userNum"));
+			String iconId=(String)multi.getFilesystemName("filename");
 			
-			
-			
+			userVO.setIconId(iconId); // 생성된 파일의 이름 가져오기
+			userVO.setUserNum(userNum);
+			String url="myPage.do?"
+					+ "selUserNum="+userNum
+					+ "&myListCtgr=board";
+			forward.setPath(url);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
 		
+		
 		UsersVO uvo = userDAO.getDBData(userVO);
 		System.out.println(uvo);
 		// reupload 시 파일삭제
-		if (uvo.getIconId() != null) {
+		if (uvo.getIconId() != null && !userVO.getIconId().equals("default.png")) {
 			//파일 삭제
 			filePath += "/"+uvo.getIconId();
 			File file = new File(filePath);	//파일 생성
-			if(file.exists()) {				//파일이 있을시 삭제 
+			if(file.exists() ) {				//파일이 있을시 삭제 
 				file.delete();
 			}
 		}
 		
 		if (userDAO.uploadImg(userVO)) {
 			System.out.println("업로드 성공");
+			HttpSession session = request.getSession();
+			session.setAttribute("user", uvo);
 			forward.setRedirect(false);
-			forward.setPath("main.do");
 		}
 		else {
 			System.out.println("업로드 실패");
@@ -72,7 +80,6 @@ public class ActionImgUpload implements Action {
 				file.delete();
 			}
 			forward.setRedirect(false);
-			forward.setPath("form.do");
 		}
 		return forward;
 	}
